requirements.txt
# XinBanAI V1 - Python 依赖清单
# ================================
# 使用方式：pip install -r requirements.txt
# 环境要求：Python 3.8+
# ================================

# 核心依赖
# -------------------------------
openai>=1.0.0          # 阿里云百炼 Qwen API 客户端（兼容模式）
faiss-cpu>=1.7.4       # Facebook AI 向量检索库（CPU 版本）
numpy>=1.20.0          # 数值计算和数组处理
pytz>=2021.1           # 时区处理支持

# 可选依赖（数据库可视化功能）
# -------------------------------
flask>=2.0.0           # Web 服务器框架（用于 db_viewer.py）
pandas>=1.3.0          # 数据处理和导出（CSV/JSON）

# 本地推理模型依赖（实验性功能）
# -------------------------------
# llama-cpp-python>=0.2.0  # Qwen2.5-7B-4bit 量化模型推理（需单独安装）
# 注意：llama-cpp-python 需要编译环境，详见 LOCAL_INFERENCE.md

# 开发工具（可选）
# -------------------------------
# pytest>=7.0.0          # 单元测试框架
# black>=22.0.0          # 代码格式化工具
# flake8>=4.0.0          # 代码风格检查
